+++ resolwe/flow/serializers/annotations.py
@@ -1,5 +1,8 @@
 """Resolwe annotations serializer."""
+from typing import Any
+
 from django.db import models
+from django.db.models import Q
 
 from rest_framework import serializers
 from rest_framework.fields import empty
@@ -94,13 +97,71 @@ class AnnotationPresetSerializer(ResolweBaseSerializer):
         fields = read_only_fields + ("name", "fields", "contributor")
 
 
+class AnnotationValueListSerializer(serializers.ListSerializer):
+    """Perform bulk update of annotation values to speed up requests."""
+
+    def create(self, validated_data: Any) -> Any:
+        """Perform efficient bulk create."""
+        return AnnotationValue.objects.bulk_create(
+            AnnotationValue(**data)
+            for data in validated_data
+            if data["_value"]["value"] is not None
+        )
+
+    def update(self, instance, validated_data: Any):
+        """Perform efficient bulk create/update/delete."""
+        # Read existing annotations in a single query.
+        query = Q()
+        for data in validated_data:
+            query |= Q(field=data["field"], entity=data["entity"])
+        existing_annotations = AnnotationValue.objects.filter(query)
+
+        # Create a mapping between (field, entity) and the existing annotations.
+        annotation_map = {
+            (annotation.field, annotation.entity): annotation
+            for annotation in existing_annotations
+        }
+
+        self.instance = []
+        to_update = []
+        to_create = []
+        to_delete = []
+        for data in validated_data:
+            if value := annotation_map.get((data["field"], data["entity"])):
+                if data["_value"]["value"] is None:
+                    to_delete.append(value.pk)
+                else:
+                    to_update.append((value, data))
+            else:
+                to_create.append(data)
+
+        # Bulk create new annotations.
+        self.instance += self.create(to_create)
+        # Update annotations.
+        for value, data in to_update:
+            self.instance.append(self.child.update(value, data))
+        # Bulk delete annotations.
+        AnnotationValue.objects.filter(pk__in=to_delete).delete()
+        return self.instance
+
+    def validate(self, attrs: Any) -> Any:
+        """Validate list of annotation values."""
+        if len(set((attr["field"], attr["entity"]) for attr in attrs)) != len(attrs):
+            raise serializers.ValidationError(
+                "Duplicate annotation values for the same entity and field."
+            )
+        return super().validate(attrs)
+
+
 class AnnotationValueSerializer(ResolweBaseSerializer):
     """Serializer for AnnotationValue objects."""
 
     def __init__(self, instance=None, data=empty, **kwargs):
         """Rewrite value -> _value."""
-        if data is not empty and "value" in data:
-            data["_value"] = {"value": data.pop("value", None)}
+        if data is not empty:
+            for entry in data if isinstance(data, list) else [data]:
+                if "value" in entry:
+                    entry["_value"] = {"value": entry.pop("value")}
         super().__init__(instance, data, **kwargs)
 
     class Meta:
@@ -111,3 +172,4 @@ class AnnotationValueSerializer(ResolweBaseSerializer):
         update_protected_fields = ("id", "entity", "field")
         fields = read_only_fields + update_protected_fields + ("value", "_value")
         extra_kwargs = {"_value": {"write_only": True}}
+        list_serializer_class = AnnotationValueListSerializer